Added a File named 2019-2020 which contains the uptodate Simulation and CAD of the UGV
The simulation that is currently in the folder Simulation called cadInput has a model of the UGV in it I believe that it can be used right now as a movable vehicle. just needs to have the code written for it. I will be adding a readme.txt so that all objects have a description of how to reference them.