Kyle's CAD world Conversion Attempt 1
I tried to make the CAD of the UGV usable through an external controller run from VS Code. I got the code to run, but if you run it, you can tell its not correct, it just dislocates the tires/motors from the chassis and spins them in the air. My hypothesis is that the issue is with the model not the code.